// src/modules/school/school.types.ts
export interface SchoolAttributes {
	id?: string;
	idSchool?: string;
	name?: string;
	size?: string;
	district?: string;
	email?: string;
	tel?: string;
	address?: string;
	junior?: string;
	senior?: string;
	director?: string;
	nTeacher?: string;
	nPersonnel?: string;
	teachingStyle?: string;
	openClass?: string;
	userId?: string;
	createdAt?: Date;
	updatedAt?: Date;
}